move overlay light dismiss in the trigger
