+++ packages/components/src/tooltip/src/Tooltip.tsx
@@ -20,26 +20,6 @@ export function InnerTooltip({
     forwardedRef,
     ...rest
 }: InnerTooltipProps) {
-    const { close } = useTooltipTriggerContext();
-
-    const [focusScope, setFocusRef] = useFocusScope();
-
-    const tooltipRef = useMergedRefs(forwardedRef, setFocusRef);
-
-    const overlayDismissProps = useOverlayLightDismiss(focusScope, {
-        hideOnEscape: true,
-        hideOnLeave: true,
-        hideOnOutsideClick: false,
-        onHide: useEventCallback((event: SyntheticEvent) => {
-            close(event);
-            // Ignore events related to the trigger.
-            // if (!isTargetParent(event.target, triggerRef) && (event as FocusEvent).relatedTarget !== triggerRef.current) {
-            //     close(event);
-            // }
-        }),
-        trigger: "hover"
-    });
-
     return (
         <Text
             {...mergeProps(
@@ -47,10 +27,9 @@ export function InnerTooltip({
                 {
                     as,
                     className: "o-ui-tooltip",
-                    ref: tooltipRef,
+                    ref: forwardedRef,
                     role: "tooltip"
-                },
-                overlayDismissProps
+                }
             )}
         >
             {children}
